requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.7.2
    # via django
certifi==2023.5.7
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
charset-normalizer==3.1.0
    # via requests
convertdate==2.4.0
    # via holidays
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
dateparser==1.1.8
    # via -r .\requirements.in
dj-rest-auth==3.0.0
    # via -r .\requirements.in
django==4.2.1
    # via
    #   -r .\requirements.in
    #   dj-rest-auth
    #   django-ckeditor
    #   django-filter
    #   django-image-optimizer
    #   django-js-asset
    #   django-tempus-dominus
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-jsonmask
    #   drf-yasg
django-ckeditor==6.5.1
    # via -r .\requirements.in
django-filter==23.1
    # via -r .\requirements.in
django-image-optimizer==1.0.3
    # via -r .\requirements.in
django-js-asset==2.0.0
    # via django-ckeditor
django-tempus-dominus==5.1.2.17
    # via -r .\requirements.in
djangorestframework==3.14.0
    # via
    #   -r .\requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-jsonmask
    #   drf-yasg
djangorestframework-simplejwt==5.2.2
    # via -r .\requirements.in
drf-jsonmask==0.3.0
    # via -r .\requirements.in
drf-yasg==1.21.5
    # via -r .\requirements.in
elastic-apm==6.15.1
    # via -r .\requirements.in
gunicorn==20.1.0
    # via -r .\requirements.in
hijri-converter==2.3.1
    # via holidays
holidays==0.23
    # via -r .\requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonmask==0.1.1
    # via drf-jsonmask
korean-lunar-calendar==0.3.1
    # via holidays
markupsafe==2.1.2
    # via jinja2
packaging==23.1
    # via drf-yasg
pillow==9.5.0
    # via
    #   -r .\requirements.in
    #   django-image-optimizer
    #   python-resize-image
psycopg[binary]==3.1.9
    # via -r .\requirements.in
psycopg-binary==3.1.9
    # via psycopg
pyjwt==2.7.0
    # via djangorestframework-simplejwt
pymeeus==0.5.12
    # via convertdate
python-dateutil==2.8.2
    # via
    #   dateparser
    #   holidays
python-decouple==3.8
    # via -r .\requirements.in
python-resize-image==1.1.20
    # via django-image-optimizer
pytz==2023.3
    # via
    #   dateparser
    #   djangorestframework
    #   drf-yasg
regex==2023.5.5
    # via dateparser
requests==2.31.0
    # via
    #   coreapi
    #   django-image-optimizer
    #   python-resize-image
    #   tinify
ruamel-yaml==0.17.30
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
sentry-sdk==1.21.1
    # via -r .\requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tinify==1.6.0
    # via django-image-optimizer
typing-extensions==4.6.2
    # via psycopg
tzdata==2023.3
    # via
    #   django
    #   psycopg
    #   tzlocal
tzlocal==5.0.1
    # via dateparser
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==2.0.2
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
wrapt==1.15.0
    # via elastic-apm

# The following packages are considered to be unsafe in a requirements file:
# setuptools
